+++ lib/bsd/ioctl.c
@@ -33,7 +33,7 @@ ioctl(int fd, unsigned long request, void* arg)
 			return -1;
 		}
 		/* this works if the file is buffered somehow */
-		*(int32_t*)arg = d.st_size;
+		*(long*)arg = d.st_size;
 		return 0;
 	} else {
 		errno = EINVAL;
