+++ lib/utf/utfrrune.c
@@ -26,9 +26,9 @@
 #include "utfdef.h"
 
 char*
-utfrrune(char *s, int32_t c)
+utfrrune(char *s, long c)
 {
-	int32_t c1;
+	long c1;
 	Rune r;
 	char *s1;
