+++ lib/v/nrand.c
@@ -12,7 +12,7 @@
 #define	MASK	0x7FFFFFFFL
 #define	FRACT	(1.0 / (MASK + 1.0))
 
-extern int32_t lrand(void);
+extern long lrand(void);
 
 double
 frand(void)
